requirements/test.txt
-r ./devel.txt

pytest==5.4.3
pytest-aiohttp==0.3.0
pytest-cov==2.10.0
tox==3.16.1